Topology: Use pipe-volume-switch-capture in nocodec topologies
This patch replaces use of pipe-volume-capture macro to
pipe-volume-switch-capture in nocodec topologies for APL, CNL, ICL,
and TGL based platforms. It allows to test of volume component
mute switch control in SSP loopback that is used by default in
nocodec topologies. The testing of mute switch feature is simplest
to do with loopback topologies. Nocodec topologies are not in
mainstream usage so this is the safest option to enable testing.

The mute switch add should be possible to all capture topologies but
it can be done later once it is confirmed it is safe to do (avoid
accidental muted audio or problems with UCM).

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>